Fix the A* heuristic.
The heuristic intended in the code was previously:
    2 * min(count of occupied rows, count of occupied columns) - 1)

It was an overestimate and would have lead to incorrect solutions if used.
For example, two items on different rows and columns could be picked up in
only two moves, given a wall that allowed an L-shaped path.

However, there was a separate bug that caused the heuristic to be calculated
as zero for all states, meaning the pathfinding devolved into Djikstra's
algorithm. Solutions were correct but slower to find than they should have
been.

Both issues have been fixed. The new heuristic is:
    min(count of occupied rows, count of occupied columns)