Fix conflicts in parallel builds.
CMake's documentation warns that there may be conflicts when parallel build
targets depend on the output of a single custom command. This did in fact
happen - building the NTSC and PAL binaries with "make -j4" resulted in a
corrupt copy of map_includes.inc because multiple commands were run in
parallel to update the same file.

CMake's suggested fix is to create a single custom target that depends on
the output of the custom command, then make the parallel targets depend on
that. This fixes the file corruption, but introduces new problems - the
binaries are not rebuilt when the source files change.

Listing both the custom target and the command output (map_includes.inc) as
separate dependencies seems to work, though it goes against CMake's guidance.