Disable AtariVox memory card support.
Writing to the memory card would be inappropriate given that there are two
versions of the game with different levels, which are currently assigned
the same address space on the memory card.

To make things worse, the way the game handles having fewer than 24 levels
is to treat any missing levels as completed, meaning the state of levels
9-24 in the cartridge version would be wiped out immedidately when loading
this version.